+++ contracts/satoshi-bridge/src/bitcoin_utils/mod.rs
@@ -1,1 +1,2 @@
 pub mod psbt_wrapper;
+pub mod transaction;
